requirements.txt
Flask==2.3.2
requests==2.31.0
openai==1.59.7
python-gitlab==3.15.0
retrying==1.3.4 